# src/main/resources/static/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Keycloak Multi-Tenant API
  description: API REST para la gestión de autenticación en un entorno multi-tenant con Keycloak. Utiliza un flujo de Password Grant Type para el login.
  version: 1.0.0
servers:
  - url: 'http://localhost:8081/api/v1'
    description: Servidor de desarrollo
paths:
...
  /{realm}/register:
    get:
      summary: Obtiene información del tenant para el registro
      operationId: getRegisterInfo
      tags:
        - User Management
      parameters:
        - name: realm
          in: path
          description: El identificador del tenant (realm)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Información del tenant para el registro
          content:
            application/json:
              schema:
                type: object
                properties:
                  realm:
                    type: string
                  registerRequest:
                    $ref: '#/components/schemas/RegisterRequest'
                  keycloakRealm:
                    type: string
        '404':
          description: Tenant no reconocido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Registra un nuevo usuario en un tenant
      operationId: registerUser
      tags:
        - User Management
      parameters:
        - name: realm
          in: path
          description: El identificador del tenant (realm)
          required: true
          schema:
            type: string
      requestBody:
        description: Datos del nuevo usuario a registrar
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered. Waiting for admin approval."
                  tenantId:
                    type: string
                  keycloakRealm:
                    type: string
        '400':
          description: Error de validación, contraseñas no coincidentes o email ya registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tenant no reconocido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: El usuario ya existe o hay un conflicto con la creación en Keycloak
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor al interactuar con Keycloak
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - confirmPassword
        - email
        - firstName
        - lastName
      properties:
        username:
          type: string
          description: El nombre de usuario deseado.
          example: "new.user"
        password:
          type: string
          description: La contraseña del usuario.
          format: password
          minLength: 8
        confirmPassword:
          type: string
          description: La confirmación de la contraseña.
          format: password
          minLength: 8
        email:
          type: string
          description: La dirección de correo electrónico del usuario.
          format: email
          example: "new.user@example.com"
        firstName:
          type: string
          description: El primer nombre del usuario.
          example: "Nuevo"
        lastName:
          type: string
          description: El apellido del usuario.
          example: "Usuario"
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        details:
          type: object
          description: Detalles adicionales del error, como errores de validación.
      example:
        timestamp: "2025-08-06T06:54:30.468Z"
        status: 400
        error: "Bad Request"
        message: "Uno o mas campos tienen errores de validacion"
        details:
          username: "El nombre de usuario no puede estar vacio"